refactor: update tests to use server modules
