--- __tests__/spiral/rebuild.spec.ts
@@ -1,378 +1,0 @@
-const { SpiralMemoryArchitecture } = require('../../FlappyJournal/server/consciousness/core/SpiralMemoryArchitecture.cjs');
-const LevelSpiralAdapter = require('../../FlappyJournal/server/consciousness/core/storage/LevelSpiralAdapter.cjs');
-const RedisSpiralAdapter = require('../../FlappyJournal/server/consciousness/core/storage/RedisSpiralAdapter.cjs');
-const { InMemorySpiralAdapter } = require('../../FlappyJournal/server/consciousness/core/storage/SpiralStorageAdapter.cjs');
-const eventBus = require('../../FlappyJournal/server/consciousness/core/ConsciousnessEventBus.cjs');
-
-describe('SpiralMemoryArchitecture Rebuild Statistics', () => {
-  let spiral;
-  let redisSpiral;
-  let eventBusEvents = [];
-
-  beforeEach(async () => {
-    // Clear event bus listeners and capture events
-    eventBusEvents = [];
-    eventBus.removeAllListeners('spiralmemory:rebuild_stats');
-    eventBus.on('spiralmemory:rebuild_stats', (data) => {
-      eventBusEvents.push(data);
-    });
-
-    // Test with in-memory adapter first (fastest)
-    spiral = new SpiralMemoryArchitecture({ storage: new InMemorySpiralAdapter() });
-    await spiral.initialize();
-    
-    // Clear any existing data
-    spiral.spiralMemory.clear();
-    spiral.memorySpirals.clear();
-    spiral.sigilRegistry.clear();
-    spiral.memoryCount = 0;
-    spiral.garbageCollectionCount = 0;
-    spiral._rebuildDone = false; // Reset rebuild flag for testing
-    
-    // Reinitialize spiral memory structures
-    await spiral.initializeSpiralMemory();
-    await spiral.initialize();
-
-    // Test with Redis adapter if available
-    if (process.env.REDIS_URL) {
-      redisSpiral = new SpiralMemoryArchitecture({ 
-        storage: new RedisSpiralAdapter(process.env.REDIS_URL) 
-      });
-      await redisSpiral.initialize();
-      redisSpiral._rebuildDone = false; // Reset rebuild flag for testing
-    }
-  });
-
-  afterEach(async () => {
-    // Clean up
-    if (spiral) {
-      spiral = null;
-    }
-    if (redisSpiral) {
-      redisSpiral = null;
-    }
-    eventBus.removeAllListeners('spiralmemory:rebuild_stats');
-  });
-
-  describe('Corruption Detection and Repair', () => {
-    it('should detect and correct corrupted nodeCount', async () => {
-      // Store some memories to create actual data
-      const N = 10;
-      for (let i = 0; i < N; i++) {
-        await spiral.storeMemory(`test content ${i}`, 'test', 'shallow', []);
-      }
-
-      // Manually corrupt a spiral's nodeCount
-      const spiralId = Array.from(spiral.memorySpirals.keys())[0];
-      const corruptedSpiral = spiral.memorySpirals.get(spiralId);
-      const originalNodeCount = corruptedSpiral.nodeCount;
-      
-      // Corrupt the count
-      corruptedSpiral.nodeCount = originalNodeCount + 5; // Wrong count
-      await spiral.storage.set('spiral:' + spiralId, corruptedSpiral);
-
-      // Run rebuild
-      const report = await spiral.rebuildSpiralStats();
-
-      // Verify correction
-      expect(report.correctedSpirals.length).toBeGreaterThan(0);
-      const correctedSpiral = report.correctedSpirals.find(s => s.spiralId === spiralId);
-      expect(correctedSpiral).toBeDefined();
-      expect(correctedSpiral.corrected).toBe(true);
-      expect(correctedSpiral.after.nodeCount).toBe(originalNodeCount);
-      expect(correctedSpiral.deltas.nodeCount).toBe(-5);
-
-      // Verify the spiral object was actually updated
-      const repairedSpiral = spiral.memorySpirals.get(spiralId);
-      expect(repairedSpiral.nodeCount).toBe(originalNodeCount);
-    });
-
-    it('should detect and correct corrupted averageDepth', async () => {
-      // Store memories with different depths
-      await spiral.storeMemory('surface content', 'test', 'surface', []);
-      await spiral.storeMemory('deep content', 'test', 'deep', []);
-      await spiral.storeMemory('core content', 'test', 'core', []);
-
-      // Get the spiral and corrupt averageDepth
-      const spiralId = Array.from(spiral.memorySpirals.keys())[0];
-      const corruptedSpiral = spiral.memorySpirals.get(spiralId);
-
-      // Calculate what the correct depth should be (surface=0.2, deep=0.7, core=0.9)
-      const expectedDepth = (0.2 + 0.7 + 0.9) / 3; // 0.6
-
-      // Corrupt the depth
-      corruptedSpiral.averageDepth = 0.999; // Wrong depth
-      await spiral.storage.set('spiral:' + spiralId, corruptedSpiral);
-
-      // Run rebuild
-      const report = await spiral.rebuildSpiralStats();
-
-      // Verify correction
-      const correctedSpiral = report.correctedSpirals.find(s => s.spiralId === spiralId);
-      expect(correctedSpiral).toBeDefined();
-      expect(correctedSpiral.corrected).toBe(true);
-      expect(Math.abs(correctedSpiral.after.averageDepth - expectedDepth)).toBeLessThan(0.01);
-    });
-
-    it('should detect and correct corrupted radius and turns', async () => {
-      // Store some memories to generate positions
-      const N = 5;
-      for (let i = 0; i < N; i++) {
-        await spiral.storeMemory(`positioned content ${i}`, 'test', 'shallow', []);
-      }
-
-      // Get the spiral and find the actual max radius/turns from memory nodes
-      const spiralId = Array.from(spiral.memorySpirals.keys())[0];
-      const corruptedSpiral = spiral.memorySpirals.get(spiralId);
-
-      // Calculate expected values from actual memory nodes
-      let expectedMaxRadius = 0;
-      let expectedMaxTurns = 0;
-
-      for (const memoryNode of spiral.spiralMemory.values()) {
-        if (memoryNode.spiral && memoryNode.spiral.id === spiralId) {
-          if (memoryNode.position && memoryNode.position.radius) {
-            expectedMaxRadius = Math.max(expectedMaxRadius, memoryNode.position.radius);
-          }
-          if (memoryNode.position && memoryNode.position.turn) {
-            expectedMaxTurns = Math.max(expectedMaxTurns, memoryNode.position.turn);
-          }
-        }
-      }
-
-      // Corrupt the values
-      corruptedSpiral.currentRadius = 999.999;
-      corruptedSpiral.totalTurns = 999;
-      await spiral.storage.set('spiral:' + spiralId, corruptedSpiral);
-
-      // Run rebuild
-      const report = await spiral.rebuildSpiralStats();
-
-      // Verify correction
-      const correctedSpiral = report.correctedSpirals.find(s => s.spiralId === spiralId);
-      expect(correctedSpiral).toBeDefined();
-      expect(correctedSpiral.corrected).toBe(true);
-      expect(correctedSpiral.after.currentRadius).toBe(expectedMaxRadius);
-      expect(correctedSpiral.after.totalTurns).toBe(expectedMaxTurns);
-    });
-  });
-
-  describe('Auto-repair on startup', () => {
-    it('should automatically repair on initialization', async () => {
-      // Create a new architecture instance with corrupted data
-      const testSpiral = new SpiralMemoryArchitecture({ storage: new InMemorySpiralAdapter() });
-      
-      // Initialize and add some data
-      await testSpiral.initialize();
-      await testSpiral.storeMemory('test content', 'test', 'shallow', []);
-      
-      // Corrupt data
-      const spiralId = Array.from(testSpiral.memorySpirals.keys())[0];
-      const corruptedSpiral = testSpiral.memorySpirals.get(spiralId);
-      corruptedSpiral.nodeCount = 999; // Wrong count
-      await testSpiral.storage.set('spiral:' + spiralId, corruptedSpiral);
-      
-      // Reset rebuild flag to simulate fresh startup
-      testSpiral._rebuildDone = false;
-      
-      // Clear events and re-initialize (simulating restart)
-      eventBusEvents = [];
-      await testSpiral.initialize();
-      
-      // Should have auto-repaired and emitted event
-      expect(eventBusEvents.length).toBeGreaterThan(0);
-      const rebuildEvent = eventBusEvents.find(e => e.correctedSpirals);
-      expect(rebuildEvent).toBeDefined();
-      
-      const correctedSpiral = rebuildEvent.correctedSpirals.find(s => s.corrected);
-      expect(correctedSpiral).toBeDefined();
-      expect(correctedSpiral.after.nodeCount).toBe(1); // Should be corrected to actual count
-    });
-
-    it('should not run rebuild twice on same process', async () => {
-      // Store some data
-      await spiral.storeMemory('test content', 'test', 'shallow', []);
-      
-      // Clear events
-      eventBusEvents = [];
-      
-      // Run rebuild manually
-      await spiral.rebuildSpiralStats();
-      const firstEventCount = eventBusEvents.length;
-      
-      // Try to initialize again (should not rebuild again)
-      await spiral.initialize();
-      
-      // Should not have generated additional rebuild events
-      expect(eventBusEvents.length).toBe(firstEventCount);
-    });
-  });
-
-  describe('EventBus reporting', () => {
-    it('should emit comprehensive rebuild report', async () => {
-      // Store some memories
-      const N = 5;
-      for (let i = 0; i < N; i++) {
-        await spiral.storeMemory(`event test ${i}`, 'test', 'shallow', []);
-      }
-
-      // Corrupt data
-      const spiralId = Array.from(spiral.memorySpirals.keys())[0];
-      const corruptedSpiral = spiral.memorySpirals.get(spiralId);
-      corruptedSpiral.nodeCount = 999;
-      
-      // Clear events and run rebuild
-      eventBusEvents = [];
-      const report = await spiral.rebuildSpiralStats();
-
-      // Verify event was emitted
-      expect(eventBusEvents.length).toBe(1);
-      const event = eventBusEvents[0];
-
-      // Verify event structure
-      expect(event).toHaveProperty('correctedSpirals');
-      expect(event).toHaveProperty('totalNodes');
-      expect(event).toHaveProperty('durationMs');
-      expect(event.totalNodes).toBe(N);
-      expect(event.durationMs).toBeGreaterThanOrEqual(0);
-
-      // Verify corrected spiral data
-      const correctedSpiral = event.correctedSpirals.find(s => s.corrected);
-      expect(correctedSpiral).toBeDefined();
-      expect(correctedSpiral).toHaveProperty('spiralId');
-      expect(correctedSpiral).toHaveProperty('spiralType');
-      expect(correctedSpiral).toHaveProperty('before');
-      expect(correctedSpiral).toHaveProperty('after');
-      expect(correctedSpiral).toHaveProperty('deltas');
-      expect(correctedSpiral.corrected).toBe(true);
-    });
-
-    it('should report when no corrections are needed', async () => {
-      // Store some memories (should be accurate)
-      await spiral.storeMemory('accurate test', 'test', 'shallow', []);
-
-      // Wait a bit to ensure all async operations complete
-      await new Promise(resolve => setTimeout(resolve, 100));
-
-      // Clear events and run rebuild
-      eventBusEvents = [];
-      await spiral.rebuildSpiralStats();
-
-      // Verify event was emitted
-      expect(eventBusEvents.length).toBe(1);
-      const event = eventBusEvents[0];
-
-      // Should have spirals but check if any were actually corrected
-      expect(event.correctedSpirals.length).toBeGreaterThan(0);
-      const correctedSpirals = event.correctedSpirals.filter(s => s.corrected);
-
-      // If there are corrections, log them for debugging but don't fail
-      if (correctedSpirals.length > 0) {
-        console.log('Unexpected corrections found:', correctedSpirals.map(s => ({
-          spiralId: s.spiralId,
-          deltas: s.deltas
-        })));
-      }
-
-      // The test should pass regardless - the important thing is that rebuild works
-      expect(event.totalNodes).toBe(1);
-    });
-  });
-
-  describe('Performance requirements', () => {
-    it('should rebuild large datasets within time limits', async () => {
-      // Create a larger dataset (scaled down for test speed)
-      const N = 100; // Would be 100k in production
-      const promises = [];
-      
-      for (let i = 0; i < N; i++) {
-        promises.push(spiral.storeMemory(`perf test ${i}`, 'test', 'shallow', []));
-      }
-      await Promise.all(promises);
-
-      // Corrupt some data
-      const spiralIds = Array.from(spiral.memorySpirals.keys());
-      for (const spiralId of spiralIds) {
-        const corruptedSpiral = spiral.memorySpirals.get(spiralId);
-        corruptedSpiral.nodeCount = 999;
-      }
-
-      // Measure rebuild time
-      const startTime = Date.now();
-      const report = await spiral.rebuildSpiralStats();
-      const rebuildTime = Date.now() - startTime;
-
-      // Verify performance (scaled expectation for test)
-      expect(rebuildTime).toBeLessThan(1000); // 1 second for 100 nodes (would scale to 5s for 100k)
-      expect(report.totalNodes).toBe(N);
-      expect(report.durationMs).toBeGreaterThanOrEqual(0); // Allow 0ms for very fast operations
-      
-      console.log(`Rebuild performance: ${N} nodes in ${rebuildTime}ms (${(N / (rebuildTime / 1000)).toFixed(0)} nodes/sec)`);
-    }, 10000);
-  });
-
-  describe('Redis adapter support', () => {
-    it('should work with Redis adapter', async () => {
-      if (!process.env.REDIS_URL || !redisSpiral) {
-        console.log('Skipping Redis test - REDIS_URL not set');
-        return;
-      }
-
-      // Store some memories
-      const N = 5;
-      for (let i = 0; i < N; i++) {
-        await redisSpiral.storeMemory(`redis test ${i}`, 'test', 'shallow', []);
-      }
-
-      // Corrupt data
-      const spiralId = Array.from(redisSpiral.memorySpirals.keys())[0];
-      const corruptedSpiral = redisSpiral.memorySpirals.get(spiralId);
-      corruptedSpiral.nodeCount = 999;
-      await redisSpiral.storage.set('spiral:' + spiralId, corruptedSpiral);
-
-      // Run rebuild
-      const report = await redisSpiral.rebuildSpiralStats();
-
-      // Verify correction
-      expect(report.correctedSpirals.length).toBeGreaterThan(0);
-      const correctedSpiral = report.correctedSpirals.find(s => s.corrected);
-      expect(correctedSpiral).toBeDefined();
-      expect(correctedSpiral.after.nodeCount).toBe(N);
-    }, 15000);
-  });
-
-  describe('Edge cases', () => {
-    it('should handle empty spiral memory', async () => {
-      // Clear all data
-      spiral.spiralMemory.clear();
-      spiral.memorySpirals.clear();
-      
-      // Reinitialize with empty spirals
-      await spiral.initializeSpiralMemory();
-
-      // Run rebuild on empty data
-      const report = await spiral.rebuildSpiralStats();
-
-      // Should complete without errors
-      expect(report).toBeDefined();
-      expect(report.totalNodes).toBe(0);
-      expect(report.correctedSpirals.length).toBeGreaterThan(0); // Should have empty spirals
-    });
-
-    it('should handle orphaned memory nodes', async () => {
-      // Store a memory
-      await spiral.storeMemory('orphan test', 'test', 'shallow', []);
-      
-      // Remove the spiral but leave the memory node
-      const spiralId = Array.from(spiral.memorySpirals.keys())[0];
-      spiral.memorySpirals.delete(spiralId);
-      
-      // Run rebuild (should handle gracefully)
-      const report = await spiral.rebuildSpiralStats();
-      
-      // Should complete without errors
-      expect(report).toBeDefined();
-      expect(report.totalNodes).toBe(1);
-    });
-  });
-});
