--- __tests__/topology/topology.spec.ts
@@ -1,7 +1,0 @@
-import { propertyTestReversibility } from '../../FlappyJournal/server/consciousness/core/math/topologyProof';
-
-describe('HyperdimensionalSpiralTopology property tests', () => {
-  it('reversible foldTraversal', () => {
-    propertyTestReversibility(7, 50); // throws if fails
-  });
-});
